Added dof_hndler_for_triangulation.f90, and included some modifications in fem_space.f90 and fem_space_types.f90 that wre not in the original version used by Javier to start the project.
Eliminated done.txt since we already have the info using git log

+++ Sources/Fem/Integration/fem_space_types.f90
@@ -109,7 +109,7 @@ contains
     implicit none
     ! Parameters
     type(fem_fixed_info),  intent(inout) :: f_info 
-    integer(4)          ,  intent(in)    :: f_type, f_order, dim_space
+    integer(ip)          ,  intent(in)    :: f_type, f_order, dim_space
     logical             ,  intent(inout) :: created
 
     created = .false.
